+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Warn, without failing, on missing static assets during static generation
 - Static generation
   - Use block buffering to prevent logging from slowing down site generation
+  - Write .nojekyll
 - CLI
   - Removed `-C` argument (orthogonal to Ema)
 - Helpers